% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{ordenaR-package}
\alias{ordenaR}
\alias{ordenaR-package}
\title{ordenaR: Ordering and visualizing species composition under a gradient}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for ordering and visualizing species composition data along ecological gradients, including methods for exploratory data analysis and graphical representation.
}
\author{
\strong{Maintainer}: Edson Nilton de Moura Silva-Júnior \email{edsonbbiologia@gmail.com} (\href{https://orcid.org/0009-0003-9116-2020}{ORCID}) (GitHub: Edbbioeco)

Authors:
\itemize{
  \item Edson Nilton de Moura Silva-Júnior \email{edsonbbiologia@gmail.com} (\href{https://orcid.org/0009-0003-9116-2020}{ORCID}) (GitHub: Edbbioeco)
}

}
\keyword{internal}
