% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bar.R
\name{order_bar}
\alias{order_bar}
\title{Build ordering plots by barplots}
\usage{
order_bar(data, gradient, species, direct = TRUE, width = NULL)
}
\arguments{
\item{data}{Dataframe with gradient and species abundance columns.}

\item{gradient}{Dataframe column with gradient data. May be column name as string ("gradient") or column dataframe id position number. only one gradient per time.}

\item{species}{Dataframe columns with species abundance data. May be columns name as string (c("species_1", "species_2)) or columns dataframe id position number.}

\item{direct}{If FALSE, function assumes gradient are an indirect gradient, with categorical data type. default is TRUE.}

\item{width}{plot bar with, in numeric values. Default is NULL.}
}
\description{
Those plots are based oh Hill (1973) reciprocal average, used to find best species ordenation to understand and visualize species abundace-composition under a gradient. Can be used to direct gradient (numeric gradient, such as temperature, humidity, altitude, and other numeric variables) and indirect gradients (categorical gradientes, such as sample units, vegetation types, and other categorical variables).

Those barplots are frequently used by ecological reseaches (Magnusson & Bacchario, 2021) to understand and visualize. Those barplots are mainly used to visualize:

- How different species are ordened along a gradient;

- How species abundance are ordered along a gradient.
}
\details{
Data input must to be a dataframe or tibble class object. Keep attention to secure data are not matrix or list class objects. Data input must contaign:

- A column contain direct gradient (numeric values) or indirect gradient (categorical values). Pay attention whether your indirect gradient are contained on row names (i.e. sample units). If it is true, turn it into a columns, such using tibble::rownames_to_column() function, previously;

- A set of Columns containing species abundance data. Pay attention such species abundance data are conttained by a single column.

To Visualize data input structure, run ordenaR::data_ordenar.
}
\examples{

# Load packages

library(ordenaR)

library(purrr)

# Importing data

data("data_ordenar", package = "ordenaR")

# Visualizing data

data_ordenar

# Ploting to a direct gradient

ordenaR::order_bar(data = data_ordenar,
                       gradient = "gradient_1",
                       species = 4:8)

# Calculating to a indirect gradient

ordenaR::order_bar(data = data_ordenar,
                       gradient = "sample_unit",
                       species = 4:8,
                       direct = FALSE)

# Loop for multiple gradients

mult_gra <- function(gradient){

  paste0("Plot to gradient: ", gradient) |>
    message()

   grad <- ordenaR::order_bar(data = data_ordenar,
                                  gradient = gradient,
                                  species = 4:8)

   return(grad)

}

gradient <- paste0("gradient_", 1:2)

purrr::map(gradient, mult_gra)

}
\references{
\href{https://www.jstor.org/stable/2258931}{Hill, M. O. (1973). Reciprocal averaging: an eigenvector method of ordination. Journal of Ecology, 61:237-249};

\href{https://www.researchgate.net/publication/362367115_Exploring_patterns_in_ecological_data_with_multivariate_analyses}{Magnusson, W. E, Bacchario, F. B. (2021). Exploring patterns in ecological data with multivariate analyses. EDUA: Editora da Universidade Federal do Amazonas. Cap 2}.
}
