% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ordenar.R
\docType{data}
\name{data_ordenar}
\alias{data_ordenar}
\title{Data for exemples}
\format{
A dataset with 25 rows and 8 columns:
\describe{
  \item{sample_unit}{Row sample units, range p-1 to p-25}
  \item{gradient_1 & gradient_2}{Environmental gradients in sample units}
  \item{species_1 to species_5}{Species abundance in sample units, for five simulated species}
  ...
}
}
\source{
Data artificially build.
}
\usage{
data_ordenar
}
\description{
Simulated data to use in ordenaR package functions.
}
\keyword{datasets}
