% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{ordenaR-package}
\alias{ordenaR}
\alias{ordenaR-package}
\title{ordenaR: Ordening ang visualizing species composition under a gradient}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

functions to ordening species order in function to a gradient, such as environmental or sample unit ggradients, to best visualization how gradients affects species abundance.
}
\author{
\strong{Maintainer}: Edson Nilton de Moura Silva-Júnior \email{edsonbbiologia@gmail.com} (\href{https://orcid.org/0009-0003-9116-2020}{ORCID}) (GitHub: Edbbioeco)

Authors:
\itemize{
  \item Edson Nilton de Moura Silva-Júnior \email{edsonbbiologia@gmail.com} (\href{https://orcid.org/0009-0003-9116-2020}{ORCID}) (GitHub: Edbbioeco)
}

}
\keyword{internal}
